camel-python2/requirements-test.txt
-r requirements-base.txt
pytest-django
model-mommy
coverage